+++ translation-demo/package.json
@@ -16,6 +16,7 @@
     "@moq/watch": "0.2.17",
     "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-popover": "^1.1.17",
     "@radix-ui/react-select": "^2.1.4",
     "@radix-ui/react-slot": "^1.1.1",
     "@svta/cml-utils": "1.5.0",
